+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Fixed
 - Issue #1106: Layer names, folder names and breadcrumbs are translated.
+- Issue #1301: CustomMenuElement now also works in secondaryMenu.
 - Issue #1312: All layers of the added WMS layers are now added in a folder structure in the layerSelection.
 - Issue #1313: Added WMS layers are now no longer initially visible.
 - Issue #1314: Added WMS are added in a folder structure in the layerSelection.
@@ -49,7 +50,7 @@ Geobasiskarten: Removed old layers Geobasiskarten (farbig, graublau, schwarzgrau
 - Issue #1309: fixed translation of periodicity.
 - Issue #1316: fixed issue where currentComponent in the secondary menu would close when clicking on the map.
 - Issue #1305, Issue #1327: fixed searchbar navigation, if used action-buttons.
-- Routing: 
+- Routing:
     - fixed style for waypoints.
     - fixed elevation profile
     - fixed error message in console when switching from 2D to 3D in TSR